# docs/contracts/openapi.online.yaml
openapi: 3.1.0
info:
  title: AvoRAG — API ONLINE
  version: "1.0.0-draft"
  summary: Contrato HTTP del modo online de AvoRAG (RAG agronómico, aguacate Hass de exportación).
  description: |
    Contrato del **modo ONLINE** de la app móvil de AvoRAG. Derivado de la arquitectura
    `docs/ARQUITECTURA_ONLINE.md` y de los endpoints reales del backend FastAPI.

    Principios vinculantes (ver §0 de la arquitectura):
    - **Seguridad determinista primero (P-1):** el `semaforo` lo decide el servidor *después* del LLM;
      ningún modelo declara `verde` por sí mismo.
    - **Trazabilidad total (P-3):** toda respuesta lleva `response_id` y las cabeceras `X-AvoRAG-*`
      con las versiones correlacionadas (api/prompt/model/corpus/norm/feed).
    - **Frescura regulatoria (P-5):** un dato regulatorio sin cruce con su feed vivo dentro de SLA
      se sirve marcado (`warnings`) y NO puede salir en `verde`.
    - **Idempotencia (P-6):** las operaciones con efectos aceptan `Idempotency-Key`.
    - **0 errores:** no es 0 abstención; es cero peligrosas, cero citas falsas, cero caducado-como-vigente.
  contact:
    name: Jhonatan del Rio
    email: jhonatandelrio9@gmail.com
  license:
    name: Proprietary — © 2026 Jhonatan del Rio, todos los derechos reservados
    url: https://example.invalid/LICENSE
servers:
  - url: https://api.avorag.app/v1
    description: Producción
  - url: https://staging.api.avorag.app/v1
    description: Staging
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
  /capabilities:
    get:
      tags: [platform]
      operationId: getCapabilities
      summary: Estado por capacidad — el cliente decide a qué modo (1-4) degradar.
      security: []
      responses:
        "200":
          description: Estado de cada subsistema online.
          headers:
            X-AvoRAG-Versions: { $ref: "#/components/headers/XAvoRAGVersions" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Capabilities" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
...
  /hitl/pending:
    get:
      tags: [platform]
      operationId: hitlPending
      summary: Cola de respuestas ROJO del tenant pendientes de revisión del agrónomo (P-2).
      responses:
        "200":
          description: Lista de consultas en cola (texto solo si audit_store_text=True).
          content:
            application/json:
              schema:
                type: object
                properties:
                  pending: { type: array, items: { $ref: "#/components/schemas/HitlItem" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /hitl/decision:
    post:
      tags: [platform]
      operationId: hitlDecision
      summary: Decisión firmada del agrónomo sobre una respuesta en cola.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HitlDecisionIn" }
      responses:
        "200":
          description: Decisión registrada (con firma de tamper-evidence).
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  query_id: { type: string, format: uuid }
                  decision: { type: string, enum: [approved, rejected, edited] }
                  signature: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: La consulta no existe o no es de tu tenant. }
        "422": { $ref: "#/components/responses/ValidationError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OAuth2/OIDC. El claim `tenant` fija el aislamiento; NO se acepta `tenant` del body en prod.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Para integraciones servidor-a-servidor (ERP, laboratorio).
...
  headers:
    XAvoRAGVersions:
      description: Versiones correlacionadas de la respuesta (P-3), JSON compacto.
      schema:
        type: string
        examples: ['{"api":"1.0.0","prompt":"2026-06-15.v8","model":"claude-sonnet","corpus":"2026-06-17.1","norm":"2026-06-17","feeds":{"ica":"2026-06-17T10:00Z"}}']
...
  schemas:
    Semaforo:
      type: string
      enum: [verde, amarillo, rojo]
      description: verde=auto-responde; amarillo=cautela/pide datos; rojo=requiere agrónomo (HITL) o rehúsa.
...
    Capabilities:
      type: object
      required: [mode_hint, subsystems]
      properties:
        mode_hint: { type: integer, enum: [1, 2, 3, 4], description: "Modo recomendado (1=online-pleno … 4=fallback-offline)." }
        subsystems:
          type: object
          properties:
            llm_fuerte: { $ref: "#/components/schemas/CapState" }
            judge: { $ref: "#/components/schemas/CapState" }
            reranker: { $ref: "#/components/schemas/CapState" }
            feed_ica: { $ref: "#/components/schemas/CapState" }
            feed_ideam: { $ref: "#/components/schemas/CapState" }
            feed_lmr_ue: { $ref: "#/components/schemas/CapState" }
            feed_eeuu: { $ref: "#/components/schemas/CapState" }
    CapState:
      type: object
      properties:
        up: { type: boolean }
        as_of: { type: string, format: date-time, nullable: true }
        stale: { type: boolean, description: "True si el feed está fuera de su SLA de frescura (prohíbe verde sobre ese dato)." }
...
    HitlItem:
      type: object
      properties:
        query_id: { type: string, format: uuid }
        tenant: { type: string }
        semaforo: { $ref: "#/components/schemas/Semaforo" }
        question: { type: string }
        answer: { type: string, description: "Hash si audit_store_text=False (P-4)." }
        review_status: { type: string }
        corpus_version: { type: string }
        created_at: { type: string, format: date-time }
    HitlDecisionIn:
      type: object
      required: [query_id, reviewer_id, decision]
      properties:
        query_id: { type: string, format: uuid }
        reviewer_id: { type: string, minLength: 1, maxLength: 64 }
        decision: { type: string, enum: [approved, rejected, edited] }
        edited_text: { type: string, maxLength: 8000 }
        notes: { type: string, maxLength: 2000 }
...
    Problem:
      type: object
      description: RFC 7807 (application/problem+json).
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, default: "about:blank" }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string, description: "Código estable de AvoRAG (p.ej. AVO-VALIDATION, AVO-RATE-LIMIT, AVO-FEED-STALE)." }
        request_id: { type: string }
  responses:
    BadRequest:
      description: Entrada inválida.
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    Unauthorized:
      description: Falta o es inválido el token/clave.
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
...
    ValidationError:
      description: Validación semántica fallida (422).
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
...
    ServiceUnavailable:
      description: Dependencia caída (BD/LLM). El cliente DEBE degradar de modo.
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }